Add trampolines for all syscalls except fork

+++ kernel/threads.c
@@ -183,8 +183,8 @@ void clone_thread (uint64_t fork_rsp) {
   fork_pid = new_tcb->thread_id;
 }
 
-int fork_get_return_val (void) {
+uint64_t fork_get_return_val (void) {
   int res = fork_pid;
   fork_pid = 0;
-  return res;
+  return (uint64_t)res;
 }
